+++ src/GmicSharpNative.h
@@ -119,7 +119,6 @@ struct GmicOptions
 {
     const char* commandLine;
     const char* customResourcePath;
-    const char* customUserPath;
     const char* hostName;
     float* progress;
     bool* abort;